backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
faiss-cpu==1.7.4
pydantic==2.10.6
python-multipart==0.0.20
torch==2.6.0
requests==2.32.3
transformers==4.48.3
safetensors==0.5.2
numpy==1.26.2
PyPDF2==3.0.1
sentencepiece==0.2.0


